Refactor and fix some test on test_router_user.py
