+++ server/backend/tests/conftest.py
@@ -43,6 +43,7 @@ pytest_plugins = [
     "fixtures.fixtures_crud_application",
     "fixtures.fixtures_router_application",
     "fixtures.fixtures_crud_user",
+    "fixtures.fixtures_router_user",
     "fixtures.fixtures_crud_role"
 ]
